LawPal-Backend/requirements.txt
flask
python-dotenv
gunicorn
sentence-transformers
chromadb
rank_bm25
groq
tiktoken
numpy
torch
flask-cors
